OSIRIS/commands/premium/notice_command.py
# commands/free/notice_command.py
import discord
import logging

log = logging.getLogger('MyBot.Commands.Notice')

async def execute(interaction, bot, args):
    """
    Displays a notice message during command execution.
    This is handled directly in the executor loop, but we define it here for completeness.
    
    Args:
        interaction (discord.Interaction): The interaction that triggered the command
        bot (commands.Bot): The bot instance
        args (dict): Command arguments, should contain 'message'
    """
    # This command is actually handled in the executor's main loop
    # This file exists for documentation and to prevent "command not found" errors
    message = args.get('message', 'No message provided')
    log.info(f"NOTICE command processed: {message}")
    # The actual display is handled in the executor cog
    return True
