OSIRIS/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiosqlite>=0.17.0
g4f>=0.1.0
requests>=2.28.0
asyncio>=3.4.3
pytz>=2022.1
psutil>=5.9.0
